+++ src/parser/block/paragraph.test.ts
@@ -56,6 +56,12 @@ describe('Unit: parser/block/paragraph', () => {
       assert.deepStrictEqual(inspect(parser('#a\nb\n#c\n[#d]')), [['<p><a class="hashtag" rel="noopener">#a</a><span class="newline"> </span>b<span class="newline"> </span><a class="hashtag" rel="noopener">#c</a><span class="newline"> </span><a href="#index:d" rel="noopener">d</a></p>'], '']);
     });
 
+    it('comment', () => {
+      assert.deepStrictEqual(inspect(parser('<# #>')), [[], '']);
+      assert.deepStrictEqual(inspect(parser('<# a #>')), [[], '']);
+      assert.deepStrictEqual(inspect(parser('<# #>a')), [['<p>a</p>'], '']);
+    });
+
   });
 
 });